Create Perfect_Square_Number_Detective.c
"This project is a program that takes a number with 3 or fewer missing digits represented by underscores. The program will determine if the number can be a perfect square. If it can, it will provide the largest possible perfect square number as the output; if not, it will output 'Not possible.'"